fix float

+++ packages/controller/src/job.rs
@@ -161,5 +161,5 @@ pub struct JobResponse {
 #[cw_serde]
 pub struct JobsResponse {
     pub jobs: Vec<Job>,
-    pub total_count: usize,
+    pub total_count: u32,
 }
